+++ src/internal/svg/baseNameSvg.tsx
@@ -1,0 +1,22 @@
+export const baseNameSvg = (
+  <svg
+    data-testid="ockbaseNameSvg"
+    role="img"
+    aria-label="ock-base-name"
+    width="16"
+    height="16"
+    viewBox="0 0 16 16"
+    xmlns="http://www.w3.org/2000/svg"
+    fill="none"
+    className="h-full w-full"
+  >
+    <path
+      d="M11.75 3.75C11.75 5.82107 10.0711 7.5 8 7.5C5.92893 7.5 4.25 5.82107 4.25 3.75C4.25 1.67893 5.92893 0 8 0C10.0711 0 11.75 1.67893 11.75 3.75Z"
+      fill="#0A0B0D"
+    />
+    <path
+      d="M10.99 8.01001C12.21 8.04001 13.3 8.80001 13.73 9.95001L16 16H0L2.27 9.95001C2.7 8.80001 3.79 8.03001 5.01 8.01001C5.05 8.04001 6.37 9.00001 8 9.00001C9.63 9.00001 10.95 8.04001 10.99 8.01001Z"
+      fill="#0A0B0D"
+    />
+  </svg>
+);