+++ src/swap/components/SwapProvider.tsx
@@ -42,6 +42,7 @@ export function SwapProvider({
   children: React.ReactNode;
   experimental: {
     useAggregator: boolean; // Whether to use a DEX aggregator. (default: true)
+    maxSlippage?: number; // Maximum acceptable slippage for a swap. (default: 10) This is as a percent, not basis points
   };
 }) {
   // Feature flags
@@ -110,6 +111,7 @@ export function SwapProvider({
           amountReference: 'from',
           from: source.token,
           to: destination.token,
+          maxSlippage: experimental.maxSlippage?.toString(),
           useAggregator,
         });
         // If request resolves to error response set the quoteError
@@ -131,13 +133,14 @@ export function SwapProvider({
         destination.setLoading(false);
       }
     },
-    [from, to, useAggregator, handleError],
+    [from, to, useAggregator, handleError, experimental.maxSlippage],
   );
 
   const handleSubmit = useCallback(
     // biome-ignore lint/complexity/noExcessiveCognitiveComplexity: TODO Refactor this component
     async function handleSubmit(
       onError?: (error: SwapError) => void,
+      onStart?: (txHash: string) => void | Promise<void>,
       onSuccess?: (txReceipt: TransactionReceipt) => void | Promise<void>,
     ) {
       if (!address || !from.token || !to.token || !from.amount) {
@@ -154,6 +157,7 @@ export function SwapProvider({
           from: from.token,
           to: to.token,
           useAggregator,
+          maxSlippage: experimental.maxSlippage?.toString(),
         });
 
         if (isSwapError(response)) {
@@ -166,6 +170,7 @@ export function SwapProvider({
           setPendingTransaction,
           setLoading,
           sendTransactionAsync,
+          onStart,
           onSuccess,
         });
 
@@ -200,6 +205,7 @@ export function SwapProvider({
       sendTransactionAsync,
       to.token,
       useAggregator,
+      experimental.maxSlippage,
     ],
   );